Moved postgres docker image creation from ras-rm-docker-dev
These image definitions originally lived in rm-docker-dev, then moved to
ras-rm-docker-dev when that work was done and now into their own repo
here.  These are not currently used as part of any ongoing process but
most RAS and RM services are dependent on the vanilla image.  The
ultimate aim is to replace this with a full database based on int-tests/post-tests.sql